+++ app/views/memes.py
@@ -40,7 +40,7 @@ async def index(request):
         {
             "template_id": str,
             "text_lines": [str],
-            "style": str,
+            "style": [str],
             "extension": str,
             "redirect": bool,
         }
@@ -266,7 +266,7 @@ async def render_image(
         status = 414
 
     elif id == "custom":
-        url = utils.urls.arg(request, None, "background", "alt")
+        url = utils.urls.arg(request.args, None, "background", "alt")
         if url:
             template = await models.Template.create(url)
             if not template.image.exists():
@@ -275,7 +275,7 @@ async def render_image(
                 if url != settings.PLACEHOLDER:
                     status = 415
 
-            style = utils.urls.arg(request, settings.DEFAULT_STYLE, "style")
+            style = utils.urls.arg(request.args, settings.DEFAULT_STYLE, "style")
             if not utils.urls.schema(style):
                 style = style.lower()
             if not await template.check(style):
@@ -298,7 +298,7 @@ async def render_image(
             if id != settings.PLACEHOLDER:
                 status = 404
 
-        style = utils.urls.arg(request, settings.DEFAULT_STYLE, "style", "alt")
+        style = utils.urls.arg(request.args, settings.DEFAULT_STYLE, "style", "alt")
         if not await template.check(style):
             if utils.urls.schema(style):
                 status = 415